requirements.txt
asgiref==3.5.2
boto3==1.24.9
botocore==1.27.9
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.0
Markdown==3.3.7
Pillow==9.1.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
